+++ seleniumtid/pageobjects/page_object.py
@@ -13,41 +13,35 @@ been supplied.
 
 import unittest
 import logging
-import abc
 
 from seleniumtid import selenium_driver
 from seleniumtid.pageelements.page_element import PageElement
 from seleniumtid.test_cases import AppiumTestCase
+from seleniumtid.utils import Utils
 
 
 class PageObject(unittest.TestCase):
-    __metaclass__ = abc.ABCMeta
+    """
+    :type driver: selenium.webdriver.remote.webdriver.WebDriver
+    :type utils: seleniumtid.utils.Utils
+    """
 
     def __init__(self, driver=None):
         self.logger = logging.getLogger(__name__)
-        self._driver = driver if driver else selenium_driver.driver
+        self.driver = driver if driver else selenium_driver.driver
+        self.utils = Utils(self.driver)
         self.config = selenium_driver.config
         self.app_strings = AppiumTestCase.app_strings
         self.init_page_elements()
-        if driver:
-            self._update_page_elements_driver()
+        self._update_page_elements()
 
-    @property
-    def driver(self):
-        """Get the Selenium driver
-         This method allows to autocomplete self.driver in IDEs
-
-        :returns: Selenium driver
-        :rtype: selenium.webdriver.remote.webdriver.WebDriver
-        """
-        return self._driver
-
-    @abc.abstractmethod
     def init_page_elements(self):
         """Method to initialize page elements"""
+        pass
 
-    def _update_page_elements_driver(self):
-        """Assign driver to all page elements of this page object"""
-        for element in self.__dict__.values():
+    def _update_page_elements(self):
+        """Copy driver and utils instances to all page elements of this page object"""
+        for element in self.__dict__.values() + self.__class__.__dict__.values():
             if isinstance(element, PageElement):
                 element.set_driver(self.driver)
+                element.set_utils(self.utils)