+++ seleniumtid/selenium_wrapper.py
@@ -25,15 +25,21 @@ class SeleniumWrapper(object):
     driver = None
     logger = None
     config = ExtendedConfigParser()
-    conf_properties_files = None
-    conf_logging_file = None
     browser_info = None
-    screenshots_path = None
-    screenshots_number = None
-    videos_path = None
-    videos_number = None
+    # Configuration and output files
+    config_directory = None
     output_directory = None
-    baseline_directory = None
+    config_properties_filenames = None
+    config_log_filename = None
+    # Screenshots configuration
+    screenshots_directory = None
+    screenshots_number = None
+    # Videos configuration
+    videos_directory = None
+    videos_number = None
+    # Visual Testing configuration
+    visual_output_directory = None
+    visual_baseline_directory = None
     visual_number = None
 
     def __new__(cls, *args, **kwargs):
@@ -42,73 +48,73 @@ class SeleniumWrapper(object):
             cls._instance = super(SeleniumWrapper, cls).__new__(cls, *args, **kwargs)
         return cls._instance
 
-    def configure_logger(self):
-        """Configure selenium instance logger"""
-        # Get logging filename from system properties
-        try:
-            conf_logging_file = os.environ['Files_logging']
-        except KeyError:
-            conf_logging_file = os.path.join('conf', 'logging.conf')
+    def get_configured_value(self, system_property_name, specific_value, default_value):
+        """Get configured value from system properties, method parameters or default value
 
-        # Get logger filename from system properties
+        :returns: configured value
+        """
         try:
-            log_filename = os.path.normpath(os.environ['Files_log_filename']).replace('\\', '\\\\')
+            return os.environ[system_property_name]
         except KeyError:
-            log_filename = 'selenium.log'
+            return specific_value if specific_value else default_value
+
+    def configure_logger(self, tc_config_log_filename=None, tc_output_log_filename=None):
+        """Configure selenium instance logger
+
+        :param tc_config_log_filename: test case specific logging config file
+        :param tc_output_log_filename: test case specific output logger file
+        """
+        # Get config and output logger files
+        config_log_filename = self.get_configured_value('Config_log_filename', tc_config_log_filename, 'logging.conf')
+        config_log_filename = os.path.join(self.config_directory, config_log_filename)
+        output_log_filename = self.get_configured_value('Output_log_filename', tc_output_log_filename, 'selenium.log')
+        output_log_filename = os.path.join(self.output_directory, output_log_filename)
+        output_log_filename = output_log_filename.replace('\\', '\\\\')
 
         # Configure logger if logging filename has changed
-        if self.conf_logging_file != conf_logging_file:
+        if self.config_log_filename != config_log_filename:
             try:
-                logging.config.fileConfig(conf_logging_file, {'logfilename': log_filename}, False)
+                logging.config.fileConfig(config_log_filename, {'logfilename': output_log_filename}, False)
             except Exception as exc:
-                print "[WARN] Error reading logging config file '{}': {}".format(conf_logging_file, exc)
-            self.conf_logging_file = conf_logging_file
+                print "[WARN] Error reading logging config file '{}': {}".format(config_log_filename, exc)
+            self.config_log_filename = config_log_filename
             self.logger = logging.getLogger(__name__)
 
-    def configure_properties(self):
-        """Configure selenium instance properties"""
-        # Get properties filename from system properties
-        try:
-            conf_properties_files = os.environ['Files_properties']
-        except KeyError:
-            conf_properties_files = os.path.join('conf', 'properties.cfg')
+    def configure_properties(self, tc_config_prop_filenames=None):
+        """Configure selenium instance properties
+
+        :param tc_config_prop_filenames: test case specific properties filenames
+        """
+        prop_filenames = self.get_configured_value('Config_prop_filenames', tc_config_prop_filenames, 'properties.cfg')
+        prop_filenames = [os.path.join(self.config_directory, filename) for filename in prop_filenames.split(';')]
+        prop_filenames = ';'.join(prop_filenames)
 
         # Configure config if properties filename has changed
-        if self.conf_properties_files != conf_properties_files:
+        if self.config_properties_filenames != prop_filenames:
             # Initialize the config object
-            self.config = ExtendedConfigParser.get_config_from_file(conf_properties_files)
-            self.conf_properties_files = conf_properties_files
+            self.config = ExtendedConfigParser.get_config_from_file(prop_filenames)
+            self.config_properties_filenames = prop_filenames
 
             # Override properties with system properties
             self.config.update_from_system_properties()
 
-    def configure(self):
-        """Configure initial selenium instance using logging and properties files for Selenium or Appium tests"""
-        # Configure logger
-        self.configure_logger()
-        # Initialize the config object
-        self.configure_properties()
-
-        # Configure folders if browser has changed
+    def configure_visual_directories(self):
+        """Configure screenshots, videos and visual directories"""
         browser_info = self.config.get('Browser', 'browser').replace('-', '_')
+
+        # Configure directories if browser has changed
         if self.browser_info != browser_info:
             self.browser_info = browser_info
 
-            # Get output path from system properties
-            try:
-                output_path = os.environ['Files_output_path']
-            except KeyError:
-                output_path = os.path.join(os.getcwd(), 'dist')
-
             # Unique screenshots and videos directories
             date = datetime.datetime.now().strftime('%Y-%m-%d_%H%M%S')
-            self.screenshots_path = os.path.join(output_path, 'screenshots', date + '_' + browser_info)
+            self.screenshots_directory = os.path.join(self.output_directory, 'screenshots', date + '_' + browser_info)
             self.screenshots_number = 1
-            self.videos_path = os.path.join(output_path, 'videos', date + '_' + browser_info)
+            self.videos_directory = os.path.join(self.output_directory, 'videos', date + '_' + browser_info)
             self.videos_number = 1
 
             # Unique visualtests directories
-            self.output_directory = os.path.join(output_path, 'visualtests', date + '_' + browser_info)
+            self.visual_output_directory = os.path.join(self.output_directory, 'visualtests', date + '_' + browser_info)
             baseline_name = self.config.get_optional('VisualTests', 'baseline_name')
             if baseline_name:
                 language = self.config.get_optional('AppiumCapabilities', 'language', '')
@@ -117,15 +123,44 @@ class SeleniumWrapper(object):
                 baseline_name = baseline_name.replace('{platformVersion}', platform_version)
             else:
                 baseline_name = browser_info
-            self.baseline_directory = os.path.join(output_path, 'visualtests', 'baseline', baseline_name)
+            self.visual_baseline_directory = os.path.join(self.output_directory, 'visualtests', 'baseline',
+                                                          baseline_name)
             self.visual_number = 1
+
+    def configure(self, is_selenium_test=True, tc_config_directory=None, tc_output_directory=None,
+                  tc_config_prop_filenames=None, tc_config_log_filename=None, tc_output_log_filename=None):
+        """Configure initial selenium instance using logging and properties files for Selenium or Appium tests
+
+        :param is_selenium_test: true if test is a selenium or appium test case
+        :param tc_config_directory: test case specific config directory
+        :param tc_output_directory: test case specific output directory
+        :param tc_config_prop_filenames: test case specific properties filenames
+        :param tc_config_log_filename: test case specific logging config filename
+        :param tc_output_log_filename: test case specific output logger filename
+        """
+        # Get config and output directories
+        self.config_directory = self.get_configured_value('Config_directory', tc_config_directory,
+                                                          os.path.join(os.getcwd(), 'conf'))
+        self.output_directory = self.get_configured_value('Output_directory', tc_output_directory,
+                                                          os.path.join(os.getcwd(), 'output'))
+        if not os.path.exists(self.output_directory):
+            os.makedirs(self.output_directory)
+
+        # Configure logger
+        self.configure_logger(tc_config_log_filename, tc_output_log_filename)
+
+        # Initialize the config object
+        self.configure_properties(tc_config_prop_filenames)
+
+        # Configure visual directories
+        if is_selenium_test:
+            self.configure_visual_directories()
 
     def connect(self):
         """Set up the selenium driver and connect to the server
 
         :returns: selenium driver
         """
-        self.configure()
         self.driver = ConfigDriver(self.config).create_driver()
         return self.driver
 
@@ -159,4 +194,4 @@ class SeleniumWrapper(object):
         :returns: true if the browser is maximizable
         """
         browser_name = self.config.get('Browser', 'browser').split('-')[0]
-        return not self.is_mobile_test() and browser_name != 'opera'
+        return not self.is_mobile_test() and browser_name != 'opera' and browser_name != 'edge'